refactor: store.tsを責務別モジュールに分割
storeTypes, storeDbUtils, storeSeatLogic, storeClockLogic,
storeGameTransforms, storeMigrationsに分離。
store.tsにはGameStore interface, InMemoryStore, PostgresStoreを残す。
