+++ server/src/storeDbUtils.ts
@@ -1,0 +1,28 @@
+export function toIsoNow(): string {
+  return new Date().toISOString();
+}
+
+export function toIso(value: string | Date): string {
+  return value instanceof Date ? value.toISOString() : new Date(value).toISOString();
+}
+
+export function toNumber(value: number | string | bigint | null | undefined): number {
+  if (typeof value === "number") {
+    return value;
+  }
+  if (typeof value === "bigint") {
+    return Number(value);
+  }
+  if (typeof value === "string") {
+    const parsed = Number(value);
+    return Number.isFinite(parsed) ? parsed : 0;
+  }
+  return 0;
+}
+
+export function splitSqlStatements(script: string): string[] {
+  return script
+    .split(/;\s*(?:\r?\n|$)/g)
+    .map((statement) => statement.trim())
+    .filter((statement) => statement.length > 0);
+}
